+++ minis_validation/cli.py
@@ -15,29 +15,36 @@ def cli(verbose):
 
 
 def _format_to_cli(doc):
-    return doc.replace(u'\n', u'\n\n')
+    return doc.replace('\n', '\n\n')
 
 
-@cli.command(short_help='Carry simulations with various minis frequencies',
-             help=_format_to_cli(simulation.run.__doc__))
+@cli.command(short_help='Carry simulations with various minis frequencies, for details see'
+                        'docstring of ``minis_validation.simulation.run`` function.')
 @click.argument('blue_config', type=click.Path(exists=True, dir_okay=False))
 @click.argument('frequencies', type=click.Path(exists=True, dir_okay=False))
 @click.argument('jobs_configs_dir', type=click.Path(exists=True, file_okay=False))
 @click.argument('output', type=click.Path(exists=False, file_okay=False))
-@click.option('-n', '--num-cells', type=int)
-@click.option('-T', '--target-file',
-              type=click.Path(exists=True, dir_okay=False), default=None)
-@click.option('-t', '--target', type=str, required=False, default=None)
-@click.option('-s', '--seed', type=int, default=None, required=False)
+@click.option('--mpi', type=bool, default=True, show_default=True,
+              help='use MPI or threads to run parallel simulations')
+@click.option('-n', '--num-cells', type=int, help='Number of cells to simulate')
+@click.option('-T', '--target-file', type=click.Path(exists=True, dir_okay=False), default=None,
+              help='Path to a target file to override "TargetFile" of ``blue-config``')
+@click.option('-t', '--target', type=str, required=False, default=None,
+              help='Override "$target" of simulations')
+@click.option('-s', '--seed', type=int, default=None, required=False,
+              help='Pseudo-random generator seed')
 @click.option('-g', '--gids', type=str, default=None, required=False,
               help='List of GIDs separated by comma to manually set required cells to simulate.'
                    'Multiple gids: --gids=345,543. Single gid: --gids=345.')
-@click.option('-d', '--duration', type=float, default=None, required=False)
-@click.option('--forward-skip', type=float, default=5000, required=False)
+@click.option('-d', '--duration', type=float, default=None, required=False,
+              help='Override duration of simulations')
+@click.option('--forward-skip', type=float, default=5000, required=False,
+              help='Override "ForwardSkip" value of simulations')
 def simulate(blue_config: str,
              frequencies: str,
              jobs_configs_dir: str,
              output: str,
+             mpi: bool,
              num_cells: int,
              target_file: str = None,
              target: str = None,
@@ -46,12 +53,22 @@ def simulate(blue_config: str,
              duration: float = None,
              forward_skip: float = None):
     # pylint: disable=too-many-arguments
-    """CLI for ``simulation.run`` function."""
+    """CLI for ``minis_validation.simulation.run`` function.
+
+    BLUE_CONFIG is a path to a BlueConfig file
+
+    FREQUENCIES is a path to a CSV file with chosen mini frequencies to simulate
+
+    JOBS_CONFIGS_DIR is a path to a folder with YAML files describing experiment conditions
+
+    OUTPUT is a path to a folder where to store results of simulations
+    """
     assert not Path(output).exists(), f'Can\'t output to existing folder: {output}'
     simulation.run(Path(blue_config),
                    Path(frequencies),
                    Path(jobs_configs_dir),
                    Path(output),
+                   mpi,
                    num_cells,
                    None if target_file is None else Path(target_file),
                    target,
@@ -61,20 +78,30 @@ def simulate(blue_config: str,
                    forward_skip)
 
 
-@cli.command(short_help='Analyze single job results',
-             help=_format_to_cli(analysis.analyze_job.__doc__))
+@cli.command(short_help='Analyze single job results, for details see docstring '
+                        'of ``minis_validation.analysis.analyze_job`` function.')
 @click.argument('job_config_file', type=click.Path(exists=True, dir_okay=False))
 @click.argument('job_traces_dir', type=click.Path(exists=True, file_okay=False))
 def analyze_job(job_config_file: str, job_traces_dir: str):
-    """CLI for ``analysis.analyze_job`` function."""
+    """CLI for ``minis_validation.analysis.analyze_job`` function.
+
+    JOB_CONFIG_FILE is a path to a job YAML config file
+
+    JOB_TRACES_DIR is a path to a folder where stored simulations of a single job config
+    """
     analysis.analyze_job(Path(job_config_file), Path(job_traces_dir))
 
 
-@cli.command(short_help='Analyze all jobs results',
-             help=_format_to_cli(analysis.analyze_jobs.__doc__))
+@cli.command(short_help='Analyze all jobs results, for details see docstring '
+                        'of ``minis_validation.analysis.analyze_jobs`` function.')
 @click.argument('jobs_configs_dir', type=click.Path(exists=True, file_okay=False))
 @click.argument('jobs_traces_dir', type=click.Path(exists=True, file_okay=False))
 def analyze_jobs(jobs_configs_dir: str, jobs_traces_dir: str):
-    """CLI for ``analysis.analyze_jobs`` function."""
+    """CLI for ``minis_validation.analysis.analyze_jobs`` function.
+
+    JOBS_CONFIGS_DIR is a path to a folder with YAML files describing experiment conditions
+
+    JOBS_TRACES_DIR is a path to a folder where stored results of all jobs simulations
+    """
     job_results = analysis.analyze_jobs(Path(jobs_configs_dir), Path(jobs_traces_dir))
     print(job_results)